[AIP-21] Move contrib.operators.sftp_to_s3_operator providers.amazon.aws.operators.sftp_to_s3
