+++ tests/providers/amazon/aws/operators/test_sftp_to_s3.py
@@ -22,9 +22,9 @@ import unittest
 import boto3
 from moto import mock_s3
 
-from airflow.contrib.operators.sftp_to_s3_operator import SFTPToS3Operator
 from airflow.models import DAG, TaskInstance
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
+from airflow.providers.amazon.aws.operators.sftp_to_s3 import SFTPToS3Operator
 from airflow.providers.ssh.hooks.ssh import SSHHook
 from airflow.providers.ssh.operators.ssh import SSHOperator
 from airflow.utils import timezone
